pyproject.toml
[tool.poetry]
name = "apib2openapi"
version = "0.1.0"
description = "Convert docs in API Blueprint .apib format to OpenAPI 3"
authors = ["Anentropic <ego@anentropic.com>"]
license = "MIT"
homepage = "https://github.com/anentropic/python-apib2openapi"
readme = "README.md"
classifiers = [
	"Development Status :: 2 - Pre-Alpha",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
]

[tool.poetry.dependencies]
python = "^3.7"
pydantic = "^1.4"
email_validator = "^1.0.5"
typing-extensions = "^3.7.4"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
ipython = "^7.13.0"
pytype = {version = "^2020.4.1", python = ">=3.6,<3.8"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
